+++ global.d.ts
@@ -7,3 +7,8 @@ declare module '*.css?inline' {
   const content: string
   export default content
 }
+
+declare module '*.html?raw' {
+  const content: string
+  export default content
+}
