+++ src/cloud-input.ts
@@ -1,16 +1,15 @@
 import $  from 'jquery'
-import { getCandidates } from './ime_engine'
+import { getCandidates } from './ime-engine'
 import ban50Png from './img/ban50.png?inline'
 import pin50Png from './img/pin50.png?inline'
 import en50Png from './img/en50.png?inline'
 import banjiao50Png from './img/banjiao50.png?inline'
 import quanjiao50Png from './img/quanjiao50.png?inline'
 import quan50Png from './img/quan50.png?inline'
-import CloudInputCss from './cloud_input.css?inline'
+import CloudInputCss from './cloud-input.css?inline'
 import { isEditableElement, updateContent } from './utils/dom'
 import { dispatchCompositionEvent, dispatchInputEvent } from './utils/event'
 import { version } from '../package.json'
-
 
 class CloudInputObj {
   candPage = 0
@@ -27,7 +26,7 @@ class CloudInputObj {
 
   method = 1
 
-  newIn: JQuery<Element> = $(document.body)
+  newIn: JQuery<Element> = $(document.activeElement || document.body)
 
   punct = 0
 
@@ -178,6 +177,9 @@ class CloudInputObj {
         if (!this.isOn || !this.newIn) {
           return
         }
+        // TODO: 支持单引号分词
+        // TODO: 限制pinyin字符数
+        // TODO: 小尺寸页面兼容
         if (this.chiMode && this.typeOn) {
           if (e.key == 'Backspace') {
             e.preventDefault()
