--- src/loader.ts
@@ -1,4 +1,0 @@
-/**
- * loader的作用是启动后拉取最新的ime代码
- * 适用于业务无感更新的场景
- */
