+++ src/views/inputView.html
@@ -1,0 +1,1 @@
+<div></div>
