+++ vite.config.js
@@ -5,7 +5,7 @@ export default defineConfig({
   build: {
     outDir: './dist',
     lib: {
-      entry: "src/cloud_input.ts",
+      entry: "src/cloud-input.ts",
       name: "ime",
       formats: ["umd"],
       fileName: () => `cloud-ime-lib.js`,
